+++ CHANGELOG.md
@@ -1,7 +1,20 @@
+# upcoming release
+
+    - grid size can be determined by user, i.e., 3x3, 3x4, 4x3, 4,4 (and higher) grids should be possible
+
+
+# v0.4
+
+* Beta release: 
+    - window manager added: if moving half or more of a window past left top/bottom or right top/bottom
+      borders of the screen, window will be automatically resized to cover respective quarter of the screen. Likewise, moving half or more of a window past bottom left/right border of the screen, window will be 
+      automatically resized to occupy left/right half of the screen.
+
 
 # v0.3
 
-* Beta release. Fixed issue: windows now stay within the boundaries of the screen.
+* Beta release: 
+    - fixed: window cannot be moved or resized beyond boundaries of screen.
 
 # v0.2